diplom/requirements.txt
pytest
pytest-django
model_bakery
django-filter
Django
djangorestframework
psycopg2-binary